feat: rename shared components with ui
