+++ src/components/home/index.ts
@@ -1,0 +1,2 @@
+export { default as Header } from './Header';
+export { default as Informations } from './Informations';
