+++ src/components/ui/SafeView.tsx
