+++ src/sandbox/components/menu/components/Separator.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/jsx-props-no-spreading */
 
-import { Box } from '$components/shared/primitives';
-import { BoxProps } from '$components/shared/primitives/box/Box';
+import { Box } from '$components/ui/primitives';
+import { BoxProps } from '$components/ui/primitives/box/Box';
 import { Colors } from '$styles/colors';
 
 type SeparatorProps = BoxProps & {
