+++ src/sandbox/components/theme/Radiuses.sandbox.tsx
@@ -1,6 +1,6 @@
 import { ScrollView } from 'react-native';
 
-import { Box, Text } from '$components/shared/primitives';
+import { Box, Text } from '$components/ui/primitives';
 import { BorderRadii } from '$styles/radius';
 import { theme } from '$styles/theme';
