+++ src/sandbox/navigation/DebugStack.tsx
@@ -2,7 +2,7 @@ import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import { Suspense } from 'react';
 
-import FallbackLoader from '$components/shared/FallbackLoader';
+import FallbackLoader from '$components/ui/FallbackLoader';
 
 import { DebugStackParamList } from './DebugStack.types';
 import * as Pages from './pages';
